src/main/java/ru/yandex/practicum/filmorate/dal/GenreRepository.java
package ru.yandex.practicum.filmorate.dal;

import org.springframework.jdbc.core.JdbcTemplate;
import org.springframework.jdbc.core.RowMapper;
import org.springframework.stereotype.Repository;
import ru.yandex.practicum.filmorate.exceptions.GenreException;
import ru.yandex.practicum.filmorate.model.Genre;

import java.util.List;

@Repository
public class GenreRepository extends BaseRepository<Genre> {
    private static final String FIND_ALL_QUERY = "SELECT * FROM genres";
    private static final String FIND_BY_ID_QUERY = "SELECT * FROM genres WHERE id = ?";
    private static final String FIND_BY_FILM_ID_QUERY = "WITH cte AS (SELECT genre_id FROM films_genre WHERE film_id = ?) SELECT * FROM genres WHERE id IN (SELECT genre_id FROM cte);";

    public GenreRepository(JdbcTemplate jdbc, RowMapper<Genre> mapper) {
        super(jdbc, mapper);
    }

    public List<Genre> findAll() {
        return findMany(FIND_ALL_QUERY);
    }

    public Genre findById(Long id) {
        return findOne(FIND_BY_ID_QUERY, id).orElseThrow(() -> new GenreException("Жанра с id=" + id + " не существует"));
    }

    public List<Genre> findByFilmId(Long filmId) {
        return findMany(FIND_BY_FILM_ID_QUERY, filmId);
    }

}
